pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-example-table"
description = "pre-commit hook to automatically update example overview table"
version = "0.3.1"
authors = [
	{ name="Loek Le Blansch", email="loek.le-blansch.pv@renesas.com" },
]
readme = "readme.md"
requires-python = ">=3.9"
license = "BSD-3-Clause"
license-files = ["license"]
dependencies = [
	"python-frontmatter",
]

[project.urls]
Homepage = "https://bitbucket.global.renesas.com/projects/lpc_cs/repos/auto-example-table"

[project.scripts]
auto-example-table = "auto_example_table.cli:main"
